[refactor] (/survey) 설문 기능 패키지 경로 수정
