+++ src/main/java/com/yoonNeun/MyDreamPartner/domain/survey/domain/entity/Type.java
@@ -1,9 +1,8 @@
-package com.yoonNeun.MyDreamPartner.domain.category.domain;
+package com.yoonNeun.MyDreamPartner.domain.survey.domain.entity;
 
-import com.yoonNeun.MyDreamPartner.domain.type.domain.Type;
-import com.yoonNeun.MyDreamPartner.domain.typecategory.domain.TypeCategory;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,19 +12,22 @@ import java.util.Set;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "category")
-public class Category {
+@Table(name = "type")
+public class Type {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer categoryId;
+    private Integer typeId;
 
     @NotBlank
-    private String categoryName;
-
+    private String typeName;
+    
     @NotBlank
-    private String summary;
+    private String content;
 
-    @OneToMany(mappedBy = "category")
+    @NotNull
+    private Integer totalScore;
+
+    @OneToMany(mappedBy = "type")
     private Set<TypeCategory> typeCategories;
 }
