+++ src/main/java/com/yoonNeun/MyDreamPartner/domain/survey/infrastructure/SurveyRepository.java
@@ -1,6 +1,6 @@
 package com.yoonNeun.MyDreamPartner.domain.survey.infrastructure;
 
-import com.yoonNeun.MyDreamPartner.domain.type.domain.Type;
+import com.yoonNeun.MyDreamPartner.domain.survey.domain.entity.Type;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SurveyRepository extends JpaRepository<Type, Integer>, SurveyRepositoryCustom {
